RAG/rag_chatbot/requirements.txt
streamlit>=1.30.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.0.0
